Fix miscellaneous compiler warnings
Fix all compiler warnings (and one type) except for the deprectation warnings
for the gdk_screen_get_monitor_* functions used in the x11 renderer.

+++ src/plugins/splash/script/script-lib-sprite.c
@@ -462,8 +462,8 @@ static void script_lib_sprite_draw_area (script_lib_display_t *display,
                 /* In that case only draw the background if the sprite doesn't
                  * cover the complete area */
                 if (position_x > x || position_y > y ||
-                    (ply_pixel_buffer_get_width (sprite->image) + position_x) < (x + width) ||
-                    (ply_pixel_buffer_get_height (sprite->image) + position_y) < (y + height))
+                    ((int)ply_pixel_buffer_get_width (sprite->image) + position_x) < (x + width) ||
+                    ((int)ply_pixel_buffer_get_height (sprite->image) + position_y) < (y + height))
                         script_lib_draw_brackground (pixel_buffer, &clip_area, data);
         } else {
                 script_lib_draw_brackground (pixel_buffer, &clip_area, data);
